+++ modules/arrow/src/geoarrow-loader.ts
@@ -1,4 +1,5 @@
-// loaders.gl, MIT license
+// loaders.gl
+// SPDX-License-Identifier: MIT
 // Copyright (c) vis.gl contributors
 
 import type {Loader, LoaderWithParser, LoaderOptions} from '@loaders.gl/loader-utils';
@@ -7,10 +8,6 @@ import type {GeoJSONTable, GeoJSONTableBatch, BinaryGeometry} from '@loaders.gl/
 import type {ArrowTable, ArrowTableBatch} from './lib/arrow-table';
 import {parseGeoArrowSync} from './parsers/parse-geoarrow-sync';
 import {parseGeoArrowInBatches} from './parsers/parse-geoarrow-in-batches';
-
-// __VERSION__ is injected by babel-plugin-version-inline
-// @ts-ignore TS2304: Cannot find name '__VERSION__'.
-const VERSION = typeof __VERSION__ !== 'undefined' ? __VERSION__ : 'latest';
 
 export type GeoArrowLoaderOptions = LoaderOptions & {
   arrow?: {