+++ modules/arrow/test/tables/convert-arrow-to-geojson-table.spec.ts
@@ -1,4 +1,5 @@
-// loaders.gl, MIT license
+// loaders.gl
+// SPDX-License-Identifier: MIT
 // Copyright (c) vis.gl contributors
 
 import test, {Test} from 'tape-promise/tape';
@@ -8,9 +9,9 @@ import {fetchFile, parse} from '@loaders.gl/core';
 import {FeatureCollection} from '@loaders.gl/schema';
 import {GeoArrowLoader} from '@loaders.gl/arrow';
 
-test('ArrowLoader#geojson-table', (t) => {
+test('ArrowLoader#shape:geojson-table', async (t) => {
   for (const testCase of GEOARROW_TEST_CASES) {
-    testConversion(t, testCase[0], testCase[1]);
+    await testConversion(t, testCase[0], testCase[1]);
   }
   t.end();
 });