+++ modules/parquet/src/lib/parsers/parse-parquet-to-columns.ts
@@ -1,4 +1,5 @@
-// loaders.gl, MIT license
+// loaders.gl
+// SPDX-License-Identifier: MIT
 // Copyright (c) vis.gl contributors
 
 import type {ColumnarTable, ColumnarTableBatch, Schema} from '@loaders.gl/schema';
@@ -11,6 +12,9 @@ import {materializeColumns} from '../../parquetjs/schema/shred';
 import {getSchemaFromParquetReader} from './get-parquet-schema';
 import {installBufferPolyfill} from '../../polyfills/buffer';
 
+/**
+ * @deprecated
+ */
 export async function parseParquetFileInColumns(
   file: ReadableFile,
   options?: ParquetLoaderOptions
@@ -26,6 +30,9 @@ export async function parseParquetFileInColumns(
   throw new Error('empty table');
 }
 
+/**
+ * @deprecated
+ */
 export async function* parseParquetFileInColumnarBatches(
   file: ReadableFile,
   options?: ParquetLoaderOptions