+++ modules/parquet/src/lib/wasm/load-wasm-node.ts
@@ -1,3 +1,7 @@
+// loaders.gl
+// SPDX-License-Identifier: MIT
+// Copyright (c) vis.gl contributors
+
 import * as wasmNode from 'parquet-wasm/node/arrow1';
 
 export async function loadWasm(wasmUrl?: string) {