+++ modules/parquet/src/workers/parquet-worker.ts
@@ -1,7 +1,8 @@
-// loaders.gl, MIT license
+// loaders.gl
+// SPDX-License-Identifier: MIT
 // Copyright (c) vis.gl contributors
 
 import {createLoaderWorker} from '@loaders.gl/loader-utils';
-import {ParquetLoader} from '../index';
+import {ParquetLoader} from '../parquet-loader';
 
 createLoaderWorker(ParquetLoader);