+++ modules/wms/test/services/wms-service.spec.ts
@@ -1,4 +1,5 @@
-// loaders.gl, MIT license
+// loaders.gl
+// SPDX-License-Identifier: MIT
 // Copyright (c) vis.gl contributors
 
 import test from 'tape-promise/tape';
@@ -150,4 +151,36 @@ test('WMSSource#fetch override', async (t) => {
     );
     t.end();
   });
+});
+
+test('WMSSource#getImage', async (t) => {
+  const wmsService = new WMSSource({url: WMS_SERVICE_URL});
+  let getMapParameters;
+
+  // @ts-ignore
+  wmsService.getMap = (parameters) => {
+    getMapParameters = parameters;
+  };
+
+  await wmsService.getImage({
+    width: 800,
+    height: 600,
+    boundingBox: [
+      [30, 70],
+      [35, 75]
+    ],
+    layers: ['oms']
+  });
+
+  t.deepEqual(
+    getMapParameters,
+    {
+      width: 800,
+      height: 600,
+      bbox: [30, 70, 35, 75],
+      layers: ['oms']
+    },
+    'boundingBox transformed to bbox'
+  );
+  t.end();
 });