+++ CHANGELOG.md
@@ -1,5 +1,48 @@
 # CHANGELOG for loaders.gl
 
+## v4.1
+
+### v4.1.0-alpha.3
+
+- feat(mvt): Add templateUrl etc to MVTSource. (#2837)
+- fix(mvt): TileJSONLoader options.maxValues default (#2835)
+- chore(deps): bump vite from 4.4.9 to 4.5.1 (#2822)
+- fix(3d-tiles): Fix implicit tileset subtree parsing (#2823)
+- fix(json): Fix row-object and array-object shape inference (#2833)
+- fix: Fix version injection (#2834)
+- fix(zip): Local file header generation implemented (#2832)
+- fix: webpack bson bundling issue (#2831)
+- chore: Start adopting SPDX license headers (#2828)
+- chore: ensure all Writers have an encode() method (#2826)
+- feat(zip): Added append and truncate functionality (#2800)
+- fix(crypto): crc32 algorithm fix (#2819)
+- fix(crypto):  Reuse of Hash instance failures (#2824)
+
+### v4.1.0-alpha.2
+
+- feat(arrow): Triangulate on worker (#2789)
+
+### v4.1.0-alpha.1
+
+- feat(zip): cd header and zip64 info generation implemented (#2792)
+- fix: wms example (#2814)
+- fix(tile-converter): trailing slash Removing from path in i3s-server (#2813)
+- docs: fgb and tilejson docs, streamline fgbloader exports (#2811)
+- chore: upgrade to vite 5 (#2812)
+- chore: typescript 5.3 (#2810)
+- fix(GeoArrow): handle tessellation error & improve mean centers (#2803)
+- chore(parquet): GeoParquetLoader, collect the parquet loader zoo in one file (#2809)
+- GLTF: inline emod function to remove loaders.gl/math dependency (#2808)
+- chore(website): polish (#2807)
+- feat(website): Add Shapefile and KML examples (#2806)
+- chore(website): minor touches (#2805)
+- docs(las): Mark LAS v1.4 as not supported (#2804)
+- feat(website): Add GeoArrow example (#2802)
+- feat(flatgeobuf): Extract arrow compatible schema (#2801)
+- chore: update package.json keywords (#2797)
+- Convert GetImageParameters to WMSGetMapParameters (#2799)
+- chore(arrow): Support WKT and WKB geoarrow encodings (v4.1) (#2798)
+
 ## v4.0
 
 ### v4.0.4
@@ -2152,7 +2195,7 @@
 - core: Fixes to `parse` (#345)
 - docs: front page update
 - core: Add error handling to `parse*` functions (#342)
-- core: Name workers for debugger (#339)
+- core: Name workers for easier debugging (#339)
 - fix bug with read property match of undefined (#340) (#341)
 - Fix tranformMatrix regression (#338)
 - Add back DracoWorkerLoader (#337)